# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A FOSS Cards Against Humanity server written in Go",
        "title": "Cards Against Humanity API",
        "contact": {
            "name": "Danny Piper (djpiper28)",
            "url": "https://github.com/djpiper28/cards-against-humanity",
            "email": "djpiper28@gmail.com"
        },
        "license": {
            "name": "GNU GPL 3",
            "url": "https://github.com/djpiper28/cards-against-humanity"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/games/create": {
            "post": {
                "description": "Creates a new game for you to connect to via websocket afterwards",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Creates a new game",
                "parameters": [
                    {
                        "description": "create settings",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.gameCreateSettings"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/main.gameCreatedResp"
                        }
                    }
                }
            }
        },
        "/games/notFull": {
            "get": {
                "description": "Returns a list of the games",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Gets all of the games that are not full",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/gameLogic.GameInfo"
                            }
                        }
                    }
                }
            }
        },
        "/res/packs": {
            "get": {
                "description": "Gets all of the card packs as a map",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "Gets all of the resource cards packs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.getPacksResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gameLogic.CardPack": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "gameLogic.GameInfo": {
            "type": "object",
            "properties": {
                "hasPassword": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "maxPlayers": {
                    "type": "integer"
                },
                "playerCount": {
                    "type": "integer"
                },
                "playingTo": {
                    "type": "integer"
                }
            }
        },
        "gameLogic.GameSettings": {
            "type": "object",
            "properties": {
                "GamePassword": {
                    "description": "Allows a game to have a password, this will be stored in plaintext like a chad\nEmpty string is no password",
                    "type": "string"
                },
                "cardPacks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gameLogic.CardPack"
                    }
                },
                "maxPlayers": {
                    "type": "integer"
                },
                "maxRounds": {
                    "description": "Game ends when this amount of rounds is reached",
                    "type": "integer"
                },
                "playingToPoints": {
                    "description": "Game ends when someone reaches this amount of points",
                    "type": "integer"
                }
            }
        },
        "main.gameCreateSettings": {
            "type": "object",
            "properties": {
                "playerName": {
                    "type": "string"
                },
                "settings": {
                    "$ref": "#/definitions/gameLogic.GameSettings"
                }
            }
        },
        "main.gameCreatedResp": {
            "type": "object",
            "properties": {
                "gameId": {
                    "type": "string"
                },
                "playerId": {
                    "type": "string"
                }
            }
        },
        "main.getPacksResp": {
            "type": "object",
            "additionalProperties": {
                "$ref": "#/definitions/gameLogic.CardPack"
            }
        }
    }
}